[external system] store 'external-system-id' attribute for imported libraries, artifacts, and facets under .idea (IDEA-264750)
If the "Store generated project files under project root" option was enabled, IDEA didn't store a flag indicating that a library, an artifact,
or a facet was imported. So this information was lost when the project was reopened, and this led to incorrect behavior (e.g.
IDEA doesn't show a warning if a user tries to edit imported configurations). Also, if the option to store generated project files
under project root was switched off, such generated files weren't always removed from .idea.

GitOrigin-RevId: 97b572b6b6ee080c8a33d6424e8c01cb7a041c25