IDEA-77691,IDEA-307158: Add setting for Paste after Copy w/o selection
This allows users who are unhappy with the introduction of IDEA-133854
to restore the old behavior when the "Paste" action always inserts the
clipboard content at the caret position, even if it is a line copied
using Copy/Cut invoked with no selection that would break the current
line into two.

There's now also a third option available that can be more convenient in
a way that it avoids breaking the line into two by trimming the pasted
line when pasting in the middle of the line. This also avoids inserting
any unwanted indentation which is likely to be in the pasted line.

The following options are now available through the Advanced Settings:

    When pasting a line copied with no selection:
      - Paste above the caret line (default)
      - Paste at the caret position
      - Trim if pasted middle line

These options correspond to and cover the following user requests:

  - IDEA-133854 Single-line copy/cut/paste similar to Visual Studio and Sublime Text
  - IDEA-307158 Add option to restore old behavior of Copy/Cut/Paste from empty selection
  - IDEA-77691 Add option to prevent inserting line break when pasting a line copied using Copy/Cut from empty selection

(cherry picked from commit 5cc7e1c6e757fc05f19db8c01c757a5cba634a51)

IJ-CR-100617

GitOrigin-RevId: 8fb0c253be14ed2dceae131b35ba69cab576c6dc