[Gradle] set 'external-system-id' attribute for imported libraries and artifacts when the project is reimported (IDEA-264750)
If the "Store generated project files under project root" option was enabled, IDEA didn't store a flag indicating that a library or an artifact was imported. This problem is now fixed (97b572b6b6), but projects created in old versions of IDEA are still affected. This change ensures that the attribute will be properly set when an existing project is reimported.

GitOrigin-RevId: eeaa0fdb9731b3eb77895355780751b9e1accec8