settings-sync: let additional files be synchronized between server and clients: IDEA-308612
Previously, if we'd want to store something in a file unknown to this IDE, that file would not be neither deserialized from zip, nor stored in the history, nor collected from history to be sent back.

Because of that we could not, for instance, change the format of the plugins.json: we could not simply change the file format because old and new IDE versions would reparse it using its own knowledge about it format; and we could not use another file because that file, unknown to the older IDE version, would be skipped.

This change allows to passthrough unknown files located inside the metainfo folder. Therefore, IDE versions which contain this commit won't be broken by future updates.

GitOrigin-RevId: b77b0ec81de54dad9e140afe651047a565ac261d