[kotlin] make KotlinLambdaMethodFilter rely on psi while checking location inside an inline lambda
Previously KotlinLambdaMethodFilter used ClassNameCalculator to detect if location was inside an inline lambda or not. This lead to various smart stepping issues when ClassNameCalculator failed to compute lambda class names correctly. This commit makes KotlinLambdaMethodFilter perform the psi comparison of a lambda body and the expression that corresponds to a location of interest.

KTIJ-23964
KTIJ-23965