[kotlin] make KotlinRequestHint continue stepping after the kotlin fake line number in case of smart stepping
Consider the example: `1.let { it }.let { it + 1 }.let { it + 2 }`. Previously when trying to smart step into `{ it + 1 }` or `{ it + 2 }` the debugger would stop in `{ it }`, because during smart stepping the debugger would cross a fake line number that marks the end of an inline lambda, and KotlinRequestHint would stop the smart stepping process.

^KTIJ-24091 Fixed