[kotlin] introduce new heuristic to detect function literal that matches a location
Previously KotlinPositionManager relied on ClassNameCalculator to detect which function literal on a particular line matches the current jdi location. This lead to issues when ClassNameCalculator failed to compute class names correctly. For example, ClassNameCalculator computes class names of local functions according to the old backend naming scheme, which doesn't work for the IR backend. However, fixing ClassNameCalculator will not solve the issue, since there is no way to understand at runtime which backend was used for the compilation. Also, if the compiler naming scheme changes in the future, ClassNameCalculator will have to be updated again and again. This commit introduces a new heuristic to detect which inlined function literal corresponds to the given jdi location. The algorithm is based on inline marker variables borders, and while it doesn't solve the ClassNameCalculator issue completely, it still improves smart stepping and lambda breakpoints in a couple of cases.

^KTIJ-23964 Fixed
^KTIJ-23965 Fixed
^KTIJ-24090 Fixed