[KTIJ-24136] Ensure resolveAllDependsOnSourceSets and resolveDeclaredDependsOnSourceSets to not contain loops
This fix was written to support a very special SourceSet setup used by kotlinx.coroutines.
This setup lead to the .dependsOn in KotlinSourceSetInfo to contain a loop which confused the
Frontend. Using the closure functions from kotlin-tooling-core instead will prevent such
loops and handle them gracefully.

Using this method also ensures more consistency between the model entities
(KotlinSourceSet, KotlinCompilation) and the result of the resolve methods.

GitOrigin-RevId: b973c9f936e8f91635b9a837747862d52412069d