settings-sync: make External Tools exportable via Export Settings
Given that External Tools contain machine-specific information (path to executable), they are not expected to be synced: IDEA-303649
At the same time, users expect them to be exportable: IDEA-309872

There is no separate 'exportable' flag for SchemeManagers, as it is for @State. As a hotfix, let's return the DEFAULT roaming type (which will enable Export Settings for this scheme manager), but keep SettingsCategory.OTHER (which will disable settings sync).

#IDEA-309872 Fixed

GitOrigin-RevId: b88c1f6ed046b37ef610c9983e553da5e1ec77b5