settings-sync: don't push migrated data if the snapshot has been deleted on server
Previously, if there were no snapshot on the server, we were pushing local data to the server. However, if the snapshot has been previously deleted by the user, we should not push it back again.

Also, in the case of an error we should not try to push to the cloud right away, because in the situation like IDEA-307436 this will result in overwriting the server data.

#IDEA-308972 Fixed
#IDEA-307436 Fixed

(cherry picked from commit 304d92f7cff5c7783781cbc1a31903503d07bae3)

GitOrigin-RevId: cdab51364d698f5356bf1242b6687a494fee8fb3