EditorSplitters: Remove split window when closing last singleton editor
Also update the `FileEditorManagerImpl.SINGLETON_EDITOR_IN_WINDOW`'s doc
string.

Singleton file editors (diff view, disasm view) are special in a way
that the IDE already tries to keep them single in their window or split
view. In other words, they occupy a specially designated place in the
layout, which, at least in case of the disasm view, we provide
automatically. This means that the default behavior of replacing the
last closing editor in a split with a random file from the sibling split
is not the expected behavior. We are better off just removing the split
when closing the sole singleton editor.

Relates:
  - CPP-9091 Debugger: Disassemble on demand

GitOrigin-RevId: ab8cc0e6a36e766cc31151e4b5a78fc54211affd