settings-sync: get hostname lazily and safe
InetAddress.getLocalHost() can throw UnknownHostException, and also it can be slow.

At the same time, we don't need a very reliable method in this place (not speaking that getLocalHost() is not the most reliable either), and the hostname is usually written into an environment variable, either to HOSTNAME on *nix, or to COMPUTERNAME on Windows.

So we compute the hostname lazily, and return value from the environment variable in cases when it is not computed yet or at all (e.g. due to UnknownHostException).

#IDEA-309227 Fixed

(cherry picked from commit 2e952893c888fa146e1b6a8572ce270f891eb54a)

IJ-CR-101155

GitOrigin-RevId: 338ee6c54493283c225f480e308f6ca42f78e31c