[UI] IDEA-269473 Track split tool window sizes independently
There was code tracking back to 2013 that used to set weights of both
tool windows in a splitter when one is resized. At first, it used to do it
only if the split is vertical, but then this check was removed in 2021.

This broke ultrawide layout handling, when split tool windows can
have different sizes because they're side-by-side and not on top of
one another.

Since both windows receive a 'resized' event when the docking area
is resized, it makes little sense to try to update both in one event.
Whatever reasoning behind this logic was, code changed too much
since then and since it's now broken, it's best to try to do it right
instead of trying to recreate whatever logic was there 10 years ago.

Now only the resized window weight is updated when it's resized.

GitOrigin-RevId: c97f12bcb01eaa1d887fa60d32e6b3c7bb68954b