[UI] IDEA-269473 Refactoring: remove setWeight() overload
This overload was confusing: it accepted a tool window, but set
the weight of its docking area. In case of a splitter with the ultrawide
layout, it would set the weight of the entire splitter.

The overload accepting an anchor is not confusing: it doesn't
pretend to operate on individual tool windows. So make it internal
(was private) and remove the other one.

GitOrigin-RevId: ee318c0cd41d89cd6b34fb494bbd1313271c7869