Move evaluation of pushed values form outer UnindexedFilesScanner to inner UnindexedFilesScanner.CollectingIterator
This is mostly to make the code easier to read and narrow visibility of local variables (pushers, pusherExs, moduleValues).

Side effect is that the code evaluating pushed values for different origins is now executed on different threads, but this was not the objective of this commit, and no noticeable performance changes is expected from this change.

GitOrigin-RevId: f12028c75743cc76df4d009feea2df275f7c00ca