[kotlin] KTIJ-23592 Deferred indexing for script dependencies sources
The idea behind this commit is to reduce indexing time during a project opening.
Performance tests showed that source roots exclusion leads to noticeable time
reduction.

How it works
We know the type of all dependencies roots and able to say which of them are
binaries and which are sources. This data is a part of `KotlinScriptLibraryEntity`.
`KotlinScriptLibraryEntity` gets a new property `indexSourceRoots` which is `false`
by default. This leads to ignoring sources during the indexing.
While navigating a code (when a binary file is opened), a user has the option to say
"Index sources". Once clicked, corresponding `KotlinScriptLibraryEntity` gets its
`indexSourceRoots` set to `true`. Project model automatically reacts to the change
and launches indexing just for the needed entities. Now we say "yes" to including
source roots.

Note, that this commit effectively reverts experimental changes made in the scope of
71daaea59bdce8f993d6a48f5112387b5eafe7af.

^KTIJ-23592 fixed

GitOrigin-RevId: f04d3f1a0e7ca669f534caebe2a1686e94c5afdc