RUBY-30046 rename containing files of primary declarations in RBS
If a declaration in RBS matches the name of the containing file, then we also want to rename the file. This is done automatically and isn't offered as optional to the user right now.

FileOperation has a "Rename" operation added. Technically renames are just moves, but the temp file system used in tests doesn't support NIO paths so adding a simpler rename operation that takes just a String makes it easier to test.

GitOrigin-RevId: a22964ee21bdd416e98006c4721e15a934f028b2