IDEA-309378 Load HiDPI image if we failed to find standard one
If pix scale = 1.0 we try to load only images with standard size, but there can be only HiDPI images available, for example, in Tips of the Day. So, in this case we need to load HiDPI image and downscale it.

(cherry picked from commit 5a79aef73a86506a953df9ce52688c7b224de79d)

GitOrigin-RevId: 3bd1836b8c4734f0b4bf058070387cad2d731394