[git] detect root for a submodule file correctly
Submodule file path has the same path as a submodule folder, but with a different isDirectory value. For such files parent repository should be used, but after 75b214b0e9c6abdf84e3b5854880bdce0278bc72 the method did not work as expected for such files since LocalFilePath#equals was used. This commit uses AbstractRepositoryManager#getRepositoryForRootQuick which allows to quickly retrieve a root for FilePath and use it for comparison instead.

GitOrigin-RevId: 6f84c383ce6130c816ddb92b37b805fef269e456