KTIJ-23034 Fix lambdaOnSecondLine and oneLineLambda tests work with INDY lambdas
The problem was that the programme stopped on the same line three times:
1. Breakpoint outside lambda
2. After two "step into" inside lambda body
3. False hit on breakpoint inside lambda body

The cause is not related with INDY, so it was not fixed here.
The cause is that "step into" stops at the very first instruction in the method,
while breakpoint stops at the first instruction in the line, which might be not equal.
In this case, the INDY method has inserted code before start of the line with
checkNotNull for argument. In non-INDY there will be no such check.

(cherry picked from commit 3cd5eb3f009eb1891371a9acf89c02e5fdedba61)

IJ-MR-102042

GitOrigin-RevId: 7072670b72f2e3ff4df9708039d6692ba6d28d77