[workspace file index] move JvmPackageRootData to Java plugin and make it public
Kotlin plugin needs to contribute its directories to PackageIndex, so we need to have a public API for that. It's better to put it to Java plugin from the beginning, since corresponding functionality will be removed from the platform as soon as we get rid of platform API which allows plugins to access it.

GitOrigin-RevId: d66adf78641383312478950bb040bdf66ffd1eb2