[UI] IDEA-298485 Don't overwrite UI defaults for non-Compact Mode
Some defaults are set by the theme, and we don't want to overwrite them
unless we're in Compact Mode. There is no reason to explicitly set defaults
anyway as they're either removed or reset by the theme every time we switch
from Compact Mode to the normal mode or vice versa.

GitOrigin-RevId: 80ea9895f0b459835ad24ea8a64108d3dbef0d42