[java] add workaround for deadlock during PsiType initialization (IDEA-309438)
In order to fix the deadlock properly, we need to remove constants of PsiPrimitiveType from PsiType, but this would break compatibility with many plugins. So as a temporary workaround, we can ensure that initialization of PsiType is triggered before its inheritors are initialized. Since many classes refer to PsiType from they static initializers, it seems that the problem may happen only if one of PsiType's inheritors is accessed in some static initializer before PsiType is accessed. I've found only 2 such places in intellij sources, and both are fixed here.

GitOrigin-RevId: 4f4ed9d363a3c317d5719112527e64d80ba256f1