[workspace model] refactoring: extract unrelated code from RootsChangeWatcher to a separate class
Code from RootsChangeWatcher which isn't related to firing 'rootsChanged' event is now extracted to a separate class FileReferenceInWorkspaceEntityUpdater. It's also updated to improve performance: lists of changed files are converted to local variables, so there is no need to use less efficient concurrent lists for them, and instances of ChangeApplier aren't created if there are no interesting changes (that happen in most of the cases).

This is needed to migrate the code which fires 'rootsChanged' and updates WorkspaceFileIndex to watch URLs of WorkspaceFileSet instances instead of watching URLs of workspace entities.

GitOrigin-RevId: e1528784cf9e5345cd6a7eee2bf53469149cc820