[UI] IDEA-311074 Compact mode: don't re-apply the same density
updateUI() is invoked in two different ways:

1. When LaF changes (ToggleCompactMode also does this), the UI
defaults are first cleared and then reset to default values before doing
anything else.

2. When updateUI() is called directly, e.g. when the font size is changed.
In this case UI defaults are NOT cleared.

In the former case it's fine to fill all values. In the latter case, however,
this breaks some values that are patched afterward. Specifically,
Tree.rowHeight is patched by normalizing and re-scaling. If we reset
it to its default value, this breaks normalization and the end result is wrong.

Fix by storing the previous density value in the UI defaults. If they are cleared
(the former case), this value will be absent and the compact mode defaults
will kick in. In the latter case, it won't kick in as the value will match the previous
one, so whatever values were set previously, will be properly patched.

Note that at this moment it's impossible for the old density to be anything else
than either the same as the current density or null. That is, it's not possible
that, for example, oldDensityName == "COMPACT" and newDensity == "DEFAULT".
This is because ToggleCompactMode always clears defaults by re-applying the
current LaF. So currently the "ui.density" values serves more like a boolean flag.

GitOrigin-RevId: 2785e174485fc4c933e31946ef1959c347461ca8