[UI] IDEA-298485 Scale VCS Log row height properly
It's a bit special in that it doesn't use either Tree.rowHeight, List.rowHeight
or Table.rowHeight. Therefore, to be able to scale it properly with font size
changes, we need to always set its default height, so patchRowHeight will
have something to work on. This should be done no matter whether we're
in Compact Mode or not, although the values should be obviously different
in these cases.

We still need to be careful not to overwrite whatever was previously scaled
(handled by the new density != old density check) and whatever could have
been set by the theme (handled by the != null check).

GitOrigin-RevId: 86ff5aa9ca8ed172f71920110a75458770d41419