[UI] IDEA-310162 Revert SelectFileAction isEnabled update logic to fix regression
PlatformDataKeys.LAST_ACTIVE_FILE_EDITOR is missing in the data context
in case the Project View is displayed in a separate window.

Maybe there's a better way to fix it while keeping the update in the BGT,
but since it's mostly an EDT action anyway, reverting to the old update logic
should do for now.

GitOrigin-RevId: f517c9b8d615dafc477330b2170419d24ac02c46