IDEA-311545 Fix incorrect position of balloons with small height and left or right pointer
The problem occurred because the bounds with shadow border insets were provided to `BalloonImpl.AbstractPosition.isOkToHavePointer` on a first check, but without shadow border insets on a second check. So, I edited the logic to adjust pointer position and check the possibility of showing the pointer using content size without the shadow border insets.

GitOrigin-RevId: e231e114db766c60de169bc00b12aa8eef751c0b