KTIJ-23034 Fix step into lambda if a line has another lambda with '}' only
The problem was that when searching for lambdas on a string,
it did not add those with only one closing bracket '}' on the string.
This is incorrect because such lambdas can have an implicit return.
This caused inconsistency when comparing the list of found lambdas
with the list of methods which have executable code on the line.

Also, I had to rewrite the search for lambdas on line
(see `getLambdaOrFunStartingOrEndingOnLineIfInside`). Original search
in `getElementsAtLineIfAny` is implemented so that it does not find
the closing bracket from the lambda at the beginning of the string.

Also, I had to correct the creation of `SourcePosition` for the found
lambda. Previously, the line set in `SourcePosition` was the first line
of the lambda body. But this is not always true if lambda ends on
current line. This has also affected output of some non-indy tests.

(cherry picked from commit 0cf3b737d2b1cdd69470e15ecfcfd85afe0b2a1d)

IJ-MR-102042

GitOrigin-RevId: b37b710eb3e4efbadd748f741e449cf58a3e06c6