[workspace file index] migrate FileIndexBase.iterateContentUnderDirectory to use WorkspaceFileIndex (IDEA-276394)
The main difficulty with this migration was efficient processing of content under excluded directories. Usually, there is no content under an excluded directory, and it would be very inefficient to iterate over its files. However, in some cases there are content roots located under it, and they must be processed. The old implementation of FileIndexBase uses DirectoryInfo instances, which contain precomputed information about such content roots, but it's hard to efficiently implement this in WorkspaceFileIndex because it updates its data structures incrementally. So the new implementation uses data from VirtualFileUrlManager, because it provides an efficient way to find registered URLs located under a given directory, and we can be sure that all content roots are registered in VirtualFileUrlManager. Required API was added to VirtualFileUrlManager.

GitOrigin-RevId: 046be576d212ebdbf7b882ed30a98b5ca7a99e22