PY-55502 When canonizing stdlib names, perform IO-heavy PythonSdkUtil.isStdLib last
PyStdlibCanonicalPathProvider is called as part of PyiTypeProvider.getReferenceType
during inspection passes. It reads skeleton headers under the read action
to figure out which skeletons belong to the stdlib, blocking the EDT.
However, this path provider is actually applicable only to a handful of
private modules from the standard library, and it makes no sense to perform
this costly check for all other qualified names. Reversing the order of checks
should greatly reduce the number of such IO operations until we get rid of them.

GitOrigin-RevId: d5c2529ed0c8f34384517adacdc89a43b6960114