IDEA-312013 don't rely on shared flow hack in FUS logger
Hack in question: https://github.com/Kotlin/kotlinx.coroutines/issues/2603#issuecomment-808859170
It was intended to ensure that the emitted value was collected, but it doesn't work with `collectLatest`.
Thankfully, `collectLatest` is not needed here, because logs are dumped once in 12/24 hours.

This change also ensures all coroutines work in correct scopes by getting rid
of `application.coroutineScope` and `project.coroutineScope` usages.

GitOrigin-RevId: 0f93cb3a59a722ee98936a93fcc1cb301e50a9f3