[platform] IDEA-261340 Create EdtBgtTreeVisitor to handle complex scenarios
Sometimes there is a visitor delegating to another visitor. When
the call to the delegate that needs BGT is surrounded by code that
needs the EDT, things get messy.

The new DelegatingEdtBgtTreeVisitor class implements the new
EdtBgtTreeVisitor interface and allows to split the visit() call
in three calls. In case the delegate requires the EDT, it's not split,
but everything is invoked in-place. But when a BGT is needed,
then the preVisit and postVisit parts are called from the EDT,
but the visit itself is invoked on a BGT.

The logic is call preVisit first, and if it returns a non-null result,
just use it and skip the rest. Otherwise, call visit() and then
pass its result to postVisit(). It is designed to fit
the TreeUtil.promiseMakeVisible use case, and it's unclear at the
moment whether this approach is flexible enough to cover other
cases as well. This is just the first experimental stage.

(cherry picked from commit eee1a2244deb2a48ad91b7a3da0d28eaa3c17eb2)

IJ-CR-102342

GitOrigin-RevId: cffb6a37fb4904518c9e213449e8a286d7d0f6a0