[kotlin] add capability to fetch coroutine async stack trace in one call
Previously, when the Kotlin debugger needed to fetch information about a coroutine stack frame, it heavily relied on JDI communication, fetching the stack frame information individually. It led to stack trace loading time being extremely slow when debugging an app on a phone, since the JDI communication with a device is very expensive. This commit adds a wrapper over the new compiler API that allows fetching a coroutine stack trace in one call.