[UI] IDEA-312156 Preserve plain text (for sorting) when adding inplace comments
When inplace comments are added to a node's presentation, the text has to be
converted into colored text. The plain (presentable text) version would be reset
to null when that happened. However, this later led to the node's toString()
(myName) presentation to become the plain version of the full colored text,
inplace comments included. Later it was used for sorting, which shouldn't be
the case, as inplace comments are not really meant for sorting.

Fix by leaving the plain text alone when appending inplace comments.
Later it's used in PresentableNodeDescriptor.apply() to set myName,
which is then used everywhere, as it's returned by toString().

GitOrigin-RevId: 6bc3c43aca91c7110e55308ab759abe71bc0720b