[kotlin] optimize location fetching for coroutine stack frames
To assign a location to a coroutine stack frame we tried to find the existing first location on a particular line. This could be slow if the stack trace was large due to expensive JDI communication. This commit removes the searching part and creates the location bound to a particular line straightforwardly.