IDEA-311645 Introduced CachingTreePath
It caches path element depth and computes it only once on demand. Should be faster, because depth used in almost every method of TreePath implementation.

This change only improves things, but not solves everything. The issue is that some of TreePaths created by JDK and compared to caching ones.

IJ-CR-102353

GitOrigin-RevId: 6051dc735af5245fb316cc16e0e71f3be848dee5