[UI] IDEA-303793 Fit popups to the screen using their middle X coordinate
When a popup is shown, it's made to fit into the screen bounds. The screen is
either determined by the parent or, as it's often the case, when the popup
doesn't have a parent, by what screen the top-left corner of the popup is.

However, sometimes the popup slightly extents beyond the edge. There
used to be a very weak safeguard against that: instead of taking (x, y)
of the top-left corner, the (x + 1, y + 1) coordinates were used.
As it turns out, sometimes it's not enough. In particular, when the popup
has its contents aligned to the invoking component (main toolbar combo
widgets, for example), the popup itself can be shifted to the left.

In this particular case, when the project widget is on the left side of the
main toolbar (no hamburger button), its popup extends beyond the
left screen bound by way more than 1 pixel.

Fix by using the middle of the popup as a reference point to determine
which screen to fit it into. We don't do the same thing to the Y coordinate,
however, as it makes more sense to keep the popup on the same screen as
its invoker, which vertically should be very close, as the alignment algorithm
only changes the X coordinate.

GitOrigin-RevId: b7498ce05313b1324990df98743e7fb38b8e9f6e