[vcs-log-graph] introduce new api for coloring commit graph
GraphColorManager allows to define colors for the whole branches or fragments, but not for individual nodes. This commit introduces a new more flexible api with GraphColorGetter which uses the id of the node itself instead of the head node. This allows to provide two different versions implementing to strategies: GraphColorGetterByHead uses head commits and delegates to the GraphColorManager, while GraphColorGeneratorByNode introduces new behaviour and allows to return color id for each individual node.

GitOrigin-RevId: c1348492fffffd5a86624708e212950cc477a3bf