[platform] IDEA-312514 Remove deprecated getRootImmediately() call
Since Project View Pane construction is performed on the EDT, it should
not do an update, as it's a potentially slow operation.

So far the only place where it's necessary to have a valid root was in
AbstractProjectViewPaneWithAsyncSupport.initTree(), to expand
the root. Seems it isn't really necessary, as it's automatically expands
anyway when the state is restored.

GitOrigin-RevId: c30669a6d39d87f6120526a8400a4c5f1e255e08