[UI] IDEA-312549, IDEA-303793 Fit X coordinate of aligned popup to screen
To fix both cases when the component is very close to the left edge of
the screen or to the right edge. By pre-fitting X to the screen bounds
in advance, we make sure that the moveToFit() call in WizardPopup.show()
won't try to fit it into the wrong screen.

GitOrigin-RevId: b5ddb13a88d55333fdf919010f84db3700485999