[newUI] IDEA-312642 Compact mode: always scale stripe icons
There was a check that if the size is equal to the default, then
scaling is not needed. With Compact Mode, it's no longer true
as it causes the icon to have an incorrect size when the IDE
was zoomed in when in Compact Mode and then restarted and
zoomed out.

GitOrigin-RevId: c87a92c2550a92063ef9280318a31b3a131ef8c0