[util] deprecate Logger.error(Object) method
It's better to provide a meaningful error message instead of just printing an object's representation to the log. Also, the presence of a method taking Object as an argument often lead to mistakes when incorrect argument is passed to it (7c9b4d6b82256, d938c224ffaaf6, 4dc00a9836ab6, e2ae9a5abee).

GitOrigin-RevId: b38ce6bddc59540830eef9fa3dec7520366ec01b