settings-sync: provide fake implementation of the GpgSigner: IDEA-308376
Although we prohibit commit signing by passing a mock GpgConfig to the commit command, a merge commit is not covered by this call, and it is not possible to pass a GpgConfig to the MergeCommand.

Therefore, we define a trivial GpgSigner instance and set it globally.

#IDEA-308376 Fixed

(cherry picked from commit fc709bff772b0c15c2eddd3a0ea02b60a2d06a47)

IJ-CR-102680

GitOrigin-RevId: 9e43e0e375d3061c4a3e541b8902233c6abd3cad