[platform] IDEA-93140 Migrate 2022.3 tool window layouts automatically
When 2023.1 is first launched, it offers to import 2022.3 (or older) settings,
if any. We can use this to migrate the old default layout.

To do this, introduce v1 and v2 properties in ToolWindowLayoutStorageManagerState,
matching those in the old format. If we get a state that only has (at least one of) these
properties, we use them to migrate.

Also, add a deprecated window.state.xml storage to the service, so it actually
picks up those old layouts.

GitOrigin-RevId: a56f84795498ca7bc14434d7de2b145f9b862275