CPP-9091: XDebuggerEPM: Use `coroutineScope` for presentation lifetime
Launch the flow collectors tracking `activeSourceKindState`,
`navigationRequestFlow`, and `gutterIconRendererState` in the
coroutine scope defining the lifetime of ExecutionPointPresentation.
The scope is cancelled automatically at the right moment because of
the use of `.transformLatest {}`.

This also eliminates the need in clunky `.zipWithNext {}`, which used
to be used to call `hideAndDispose()` on the previous presentation.
The ExecutionPointPresentation is AutoCloseable now, and it is closed
when the coroutine scope is cancelled by the next element in the flow.

GitOrigin-RevId: 9966080e19c1555ad0dcabc62899321454e8c6b7