[kotlin] KTIJ-24494 Error incorrectly reported inside `artifacts` and `attributes` blocks of a custom Gradle plugin
The issue happened due to inability to handle Gradle's `@HasImplicitReceiver` annotation.

Information about the compiler plugin (responsible for that) is stored is a Kotlin Facet.
To be created the facet requires the compiler version. And it couldn't be detected...
because Kotlin Gradle Plugin used in the user's project was a custom one with a custom
version naming scheme: `1.8.0-space-385`. Instead of `space` the position is reserved
for a restricted set of `release`, `snapshot`, `rc`, etc. keywords.

This commit prevents swallowing of the compiler version detection problem.
Now it's reported.

^KTIJ-24494 fixed

GitOrigin-RevId: dfe4f5e03f841cdf8bb5644847ce76725afcbdc6