[tabs] Fix unpinning tab when navigating to it
It is important for FileEditorOpenOptions to have the third state for pin field. This third state means to use current pinned state if the tab is already opened. So, the pinned state will not be overridden by 'false' value if client didn't check that file is already opened and pinned.
Fix IDEA-312373 and IDEA-312047.

GitOrigin-RevId: a19be661e0a59aafd3b77dfe203366986f75af05