IFT-442 Do not reopen learning project after restart
The current opened project is updated inside `RecentProjectsManagerBase.setActivationTimestamp` on every frame activation. So, it is impossible to remove the project from recent paths on project opening, it will be added again. Also, removing it from the recent paths during project closing (before the IDE shutdown) do not work for some reason. So, I decided to create a proper way of hiding projects in the platform.

GitOrigin-RevId: c4060a163e9b57f71eb8349348ba04af6fe80e29