[platform] IDEA-313210 Introduce PathElementIdProvider for TreeState.calcId
To fix IDEA-313210 and keep IDEA-29734 and IDEA-308894 fixed,
we need a flexible mechanism to let nodes decide what is a unique
string ID that can be used to save/restore tree states.

Introduce PathElementIdProvider and let nodes implement it in cases when
a simple toString() won't do.

Remove the ugly hack that was previously used to distinguish directories
by their location strings, as it turns out that location strings may be
overridden by 3rd party plugins.

GitOrigin-RevId: b6cd56eb1f0aab10eee30e7b30e1706059b0674e