[IDE.KDoc] KTIJ-24576 Fixed linking to extension functions with bound generic type arguments
As described in the issue on YT, linking from KDoc to a function like `fun <T> List<T>.test1() {}` is possible while `fun <T : Any> List<T>.test2() {}` is not. This is presumably because a type check happens in `TCallable.substituteExtensionIfCallable` that filters out mismatches in type arguments making `/** [List.test2] */` unable to link to the function (since their generic types mismatch). However, there's no way to specify generic types in KDoc links, so we can safely ignore them using the new `ignoreTypeParameters` boolean parameter.
This parameter is then used by the `ExtensionScope` in resolveKDocLink.kt when querying extensions to link to.

GitOrigin-RevId: ec7e3b2094de46faa4d79a8de0b58354b52b92e5