don't compile same modules several times
Of course, it is not a performance issue, as we reuse already compiled sources or use incremental compilation. The change is done to simplify code and clarify, avoiding the question of why the same module is specified for compilation several times.

(cherry picked from commit 2a27fafd16b2285eb9ce0787ee163276b448ef33)

IJ-CR-102994

GitOrigin-RevId: 7527f67276aeac5238eed71a6b248fb36f9f9c49