[platform] IDEA-313096 Make ProjectViewImpl.getSelectInTargets() thread-safe
This is another tricky change, as mySelectInTargets was an ordered map,
so we had to keep the order where it matters when converting to ConcurrentHashMap.
Thankfully, the only such place is the getSelectInTargets() method itself.
We had to store the pane weights alongside the targets to order them properly
whenever an ordered list is requested.

GitOrigin-RevId: 4c099e5712adc43abf4121a428c2cc7be8b4a4b3