[platform] IDEA-313096 Make ScopeViewPane.myFilters thread-safe
It's needed to make SelectInTargetActionGroup.getChildren() thread-safe,
as it's called through ProjectViewSelectInTarget.getSubTargets().

It was already effectively immutable, so this was easy. Make it truly
immutable with Collections.unmodifiableMap() and put it into
an AtomicReference. A volatile field would do just fine, but this is
clearer in access sites and prevents accidental sequential accesses in
a non-thread-safe manner.

GitOrigin-RevId: b71dc6fd04f874c7d8be7e8c1f554471c7efdd71