[platform] IDEA-313096 Make ProjectViewImpl.getProjectViewPaneById() thread-safe
It's needed to make SelectInTargetActionGroup.getChildren() thread-safe,
as it's called through ProjectViewSelectInTarget.getSubTargets().

This one was tricky, as it's sometimes called before the panes are initialized.
To handle this, there is an isInitialized check in addProjectPane(), so it can be
added without actually initializing its Swing component. However, in exotic
cases it's possible that the panes were already initialized and then an uninitialized
pane was added. As far as I can tell, it can only happen in tests, but there was
a case before when it actually happened in production (IDEA-200322). To avoid
a possible regression, we now allow loading panes from a BGT, but if it actually
happens that we need to initialize the pane's Swing component, we defer it
with invokeLater() to be absolutely safe.

The myId2Pane field was converted from a LinkedHashMap to a ConcurrentHashMap,
but that's fine, as it didn't maintain any meaningful order anyway, as it's populated
in doAddPane(), which is called from doAddUninitializedPanes() in a loop over
myUninitializedPanes, which in turn wasn't an ordered set to begin with.

GitOrigin-RevId: 5f20614103305967f6941f455cc7bfddb8dde5e6