IDEA-303419 Fix blurry text of the editor tab if it is not selected and not hovered
Antialiasing is not working properly when there is the transparency. So I decided to simulate transparency for text color using color blending. The file icon will still be painted using transparency.
The color blending process requires the parameter, so I created the new key "EditorTabs.unselectedBlend".
Also, I found that there was the "EditorTabs.hoverAlpha" parameter, that was not mentioned in theme metadata. I decided to rename it to "EditorTabs.unselectedAlpha" and add to metadata.

GitOrigin-RevId: 23aa100e61b06969d68e111b4d005d0fbc9447d9