IDEA-310341: Respect .editorconfig on reformat with optimize imports
When the ReformatCodeAction with optimize import on one file is executed, the
optimize import processor will be run concurrently with
`notifyCodeStyleSettingsChanged()`. Which triggers a reload of the code
style settings. During that time `JavaCodeStyleSettings.getInstance(file)`
for the file will yield the project wide code style settings
because `myComputationLock` in CodeStyleCachedValueProvider#tryGetSettings
is held by the update task and the read returns null. This is a problem with
code style systems that support different formats for same files, eg.
EditorConfig.

This commit changes this to run `notifyCodeStyleSettingsChanged` after
the processor is done.

Note:
Writing a test for this is hard, because the reformat processor is only run
concurrently when non headless. So in tests it always run correctly.